+++ src/main/java/com/wanted/needswork/repository/EmployerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import javax.swing.text.html.Option;
 import java.math.BigInteger;
 
-public interface EmployerRepository extends JpaRepository <Employer,Integer> {
+public interface EmployerRepository extends JpaRepository <Employer,BigInteger> {
     Employer findByUserId(BigInteger userId);
 
     Employer findByInn(BigInteger inn);
